# backend/openapi.yaml
openapi: 3.0.0
info:
  title: ProjectPlanner API
  version: 1.0.0
servers:
  - url: http://localhost:1337
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    error_response:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    sign_up_request:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - name
        - email
        - password
    sign_up_response:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/user"
        verification_expires_at:
          type: string
          format: date-time
      required:
        - user
        - verification_expires_at
    log_in_request:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    log_in_response:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/user"
      required:
        - token
        - user
    forgot_password_request:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    forgot_password_response:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    reset_password_request:
      type: object
      properties:
        token:
          type: string
        new_password:
          type: string
          minLength: 8
      required:
        - token
        - new_password
    reset_password_response:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    verify_email_request:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    verify_email_response:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    invite_accept_request:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    invite_accept_response:
      type: object
      properties:
        workspace_id:
          type: string
        joined_at:
          type: string
          format: date-time
      required:
        - workspace_id
        - joined_at
    user:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        avatar_url:
          type: string
          nullable: true
        notify_in_app:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - notify_in_app
        - created_at
        - updated_at
...
    email_change_request:
      type: object
      properties:
        new_email:
          type: string
          format: email
      required:
        - new_email
    email_change_request_confirm:
      type: object
      properties:
        token:
          type: string
      required:
        - token
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error_response"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error_response"
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error_response"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error_response"
paths:
  /api/auth/sign_up:
    post:
      summary: Register a new user and send email verification
      operationId: sign_up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/sign_up_request"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/sign_up_response"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/auth/log_in:
    post:
      summary: Authenticate user and return JWT token
      operationId: log_in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/log_in_request"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/log_in_response"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/forgot_password:
    post:
      summary: Request password reset email
      operationId: forgot_password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/forgot_password_request"
      responses:
        "200":
          description: Password reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/forgot_password_response"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/auth/reset_password:
    post:
      summary: Reset user password using token
      operationId: reset_password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/reset_password_request"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/reset_password_response"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/auth/verify_email:
    post:
      summary: Verify user email using token
      operationId: verify_email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/verify_email_request"
      responses:
        "200":
          description: Email verification successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/verify_email_response"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/auth/invite_accept:
    post:
      summary: Accept workspace invitation
      operationId: invite_accept
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/invite_accept_request"
      responses:
        "200":
          description: Invitation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/invite_accept_response"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/email_change_requests:
    post:
      summary: Initiate email change request
      operationId: request_email_change
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/email_change_request"
      responses:
        "201":
          description: Email change request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  request_id:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
                required:
                  - request_id
                  - expires_at
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/email_change_requests/confirm:
    post:
      summary: Confirm email change with token
      operationId: confirm_email_change
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/email_change_request_confirm"
      responses:
        "200":
          description: Email changed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
security:
  - bearerAuth: []